+++ lib/new_relic.ex
@@ -53,7 +53,7 @@ defmodule NewRelic do
   defdelegate incr_attributes(attrs), to: NewRelic.Transaction.Reporter
 
   @doc """
-  Start a new "Other" Transaction.
+  Start an "Other" Transaction.
 
   This will begin monitoring the current process as an "Other" Transaction
   (ie: Not a "Web" Transaction). The first argument will be considered
@@ -90,7 +90,22 @@ defmodule NewRelic do
   defdelegate stop_transaction(), to: NewRelic.OtherTransaction
 
   @doc """
-  Call within a Transaction to prevent it from reporting.
+  Define an "Other" transaction within the given block. The return value of
+  the block is returned.
+
+  See `start_transaction` and `stop_transaction` for more details.
+  """
+  defmacro other_transaction(category, name, do: block) do
+    quote do
+      NewRelic.start_transaction(unquote(category), unquote(name))
+      res = unquote(block)
+      NewRelic.stop_transaction()
+      res
+    end
+  end
+
+  @doc """
+  Call within a transaction to prevent it from reporting.
 
   ```elixir
   def index(conn, _) do