+++ lib/new_relic/enabled_supervisor.ex
@@ -7,7 +7,7 @@ defmodule NewRelic.EnabledSupervisor do
   @moduledoc false
 
   def start_link(_) do
-    Supervisor.start_link(__MODULE__, :ok)
+    Supervisor.start_link(__MODULE__, :ok, name: __MODULE__)
   end
 
   def init(:ok) do